examples/text2text/pyproject.toml
[project]
name = "text2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.52.0",
    "pillow>=12.0.0",
]
